Bug fix - changed all times and distances to ints, from doubles

+++ assn2/TrackingServerJava/Athlete.java
@@ -7,22 +7,22 @@ public class Athlete extends java.util.Observable {
     private String last_name;
     private int age;
     private String gender;
-    private Double update_time;
-    private Double start_time;
-    private Double end_time;
+    private int update_time;
+    private int start_time;
+    private int end_time;
     private String status;
-    private double distance;
+    private int distance;
 
 
-    Athlete(int bib, Double time, String first_name, String last_name, String gender, int age){
+    Athlete(int bib, int time, String first_name, String last_name, String gender, int age){
         this.bib = bib;
         this.update_time = time;
         this.first_name = first_name;
         this.last_name = last_name;
         this.age = age;
         this.gender = gender;
-        this.start_time = 0.0;
-        this.end_time = 0.0;
+        this.start_time = 0;
+        this.end_time = 0;
         this.status = "Registered";
         this.distance = 0;
     }
@@ -31,25 +31,25 @@ public class Athlete extends java.util.Observable {
         return bib;
     }
 
-    public Double getUpdateTime() { return update_time; }
+    public int getUpdateTime() { return update_time; }
 
-    public void setUpdateTime(Double time){
+    public void setUpdateTime(int time){
         this.update_time = time;
         setChanged();
         notifyObservers();
     }
 
-    public Double getStartTime() { return start_time; }
+    public int getStartTime() { return start_time; }
 
-    public void setStartTime(Double time) {
+    public void setStartTime(int time) {
         this.start_time = time;
         setChanged();
         notifyObservers();
     }
 
-    public Double getEndTime() { return end_time; }
+    public int getEndTime() { return end_time; }
 
-    public void setEndTime(double time){
+    public void setEndTime(int time){
         this.end_time = time;
         setChanged();
         notifyObservers();
@@ -71,9 +71,9 @@ public class Athlete extends java.util.Observable {
         return age;
     }
 
-    public Double getDistance() { return this.distance; }
+    public int getDistance() { return this.distance; }
 
-    public void setDistance(Double distance){
+    public void setDistance(int distance){
         this.distance = distance;
         setChanged();
         notifyObservers();
